More refactoring of the Command object to allow it to be used in both a Command line and Web server context. - using dino.command.AbstractCommandEnvironment to contain all logging configuration, and output to the user and configuration to the command. - parsing of arguments is now optional (makes less sense in web context?)
Added probe controller to handle incoming updates of yaml information.
Uses a HTTP PUT with a script to handle the PUT

Updated logging organization to better handle using of logging for capturing output.
Many of the commands use logging as a main form of output

Fixed some issues with changeset where a changeset was committed, even when no objects had changed.
- Now the changeset object is only added to the session at the time of flush when something with revisions is_modified, added or deleted and will change on flush.

git-svn-id: https://svn.metaweb.com/svn/se-packages/dino/trunk@84470 0f5345e7-1016-0410-8812-e8f8d52c936e